move annotation parsing from c/schema-reader to binding/annotation.go and call it from go/ast-reader as well
No changes to the algorithm in this commit.
